+++ src/widgets/image_widget.rs
@@ -31,8 +31,8 @@ use std::any::Any;
 use std::collections::HashMap;
 use std::path::Path;
 
-/// This is the storage object for the `TextWidget`.  It stores the config, properties, callback registry,
-/// the font name, style, size, justification, and text message.
+/// This is the storage object for the `ImageWidget`.  It stores the config, properties, callback registry,
+/// the image name, and a scale flag.
 pub struct ImageWidget {
     config: WidgetConfig,
     system_properties: HashMap<i32, String>,
@@ -133,7 +133,7 @@ impl Widget for ImageWidget {
     /// Responds to a screen redraw only if the `CONFIG_IMAGE_POSITION` key was changed.
     fn on_config_changed(&mut self, _k: u8, _v: Config) {
         if _k == CONFIG_IMAGE_POSITION {
-            self.get_config().set_invalidate(true);
+            self.get_config().set_invalidated(true);
         }
     }