+++ src/widgets/push_button_widget.rs
@@ -44,6 +44,7 @@ pub struct PushButtonWidget {
     text_widget: TextWidget,
     active: bool,
     in_bounds: bool,
+    originated: bool,
     on_click: OnClickCallbackType,
 }
 
@@ -81,6 +82,7 @@ impl PushButtonWidget {
             text_widget,
             active: false,
             in_bounds: false,
+            originated: false,
             on_click: None,
         }
     }
@@ -92,7 +94,7 @@ impl PushButtonWidget {
             .set_color(CONFIG_COLOR_TEXT, Color::RGB(255, 255, 255));
         self.text_widget
             .set_color(CONFIG_COLOR_BASE, Color::RGB(0, 0, 0));
-        self.get_config().set_invalidate(true);
+        self.get_config().set_invalidated(true);
     }
 
     fn draw_unhovered(&mut self) {
@@ -102,7 +104,7 @@ impl PushButtonWidget {
             .set_color(CONFIG_COLOR_TEXT, Color::RGB(0, 0, 0));
         self.text_widget
             .set_color(CONFIG_COLOR_BASE, Color::RGB(255, 255, 255));
-        self.get_config().set_invalidate(true);
+        self.get_config().set_invalidated(true);
     }
 
     /// Assigns the callback closure that will be used when a button click is triggered.
@@ -173,17 +175,20 @@ impl Widget for PushButtonWidget {
             if _state {
                 self.draw_hovered();
                 self.active = true;
+                self.originated = true;
             } else {
                 let had_bounds = self.active;
 
                 self.draw_unhovered();
                 self.active = false;
 
-                if self.in_bounds && had_bounds {
+                if self.in_bounds && had_bounds && self.originated {
                     // Callback here
                     eprintln!("Call callback here: clicks={}", _clicks);
                     self.call_click_callback(_widgets, _layouts);
                 }
+
+                self.originated = false;
             }
         }